Overkill Changes, Untested

+++ src/permissions/mod.rs
@@ -1,29 +1,35 @@
 use oracle::pool::Pool;
-use oracle::Result;
-
 
 use self::structs::Permissions;
 
 pub mod structs;
 
-pub fn get_user_permissions(user_id: &str, pool: &Pool) -> Result<Permissions> {
-    let conn = pool.get()?;
-    let mut stmt = conn
-        .statement("SELECT PERMISSION FROM ODBC_JHC.PERMISSIONS_JHC WHERE USERNAME = :user_id")
-        .build()?;
-    let rows = stmt.query(&[&user_id])?;
+pub fn get_user_permissions(user_id: &str, pool: &Pool) -> Result<Permissions, oracle::Error> {
+    let conn = pool.get();
+    if conn.is_err() {
+        error!("Error connecting to DB");
+        return Err(conn.err().unwrap());
+    }
+    let conn = conn.unwrap();
 
-    let mut permission: Permissions = Permissions {
-        users: Some(false),
-        permissions: Some(false),
-        query: Some(false),
-        images: Some(false),
-        cost: Some(false),
-        admin: Some(false),
-        stock: Some(false),
-        reports: Some(false),
-        stores: Some(false),
-    };
+    let stmt = conn
+        .statement("SELECT PERMISSION FROM ODBC_JHC.PERMISSIONS_JHC WHERE USERNAME = :user_id")
+        .build();
+    if stmt.is_err() {
+        error!("Error building statement");
+        return Err(stmt.err().unwrap());
+    }
+    let mut stmt = stmt.unwrap();
+
+    let rows = stmt.query(&[&user_id]);
+    if rows.is_err() {
+        error!("Error executing query");
+        return Err(rows.err().unwrap());
+    }
+    let rows = rows.unwrap();
+
+    let mut permission: Permissions = Permissions::new();
+
     for row_result in rows {
         let row = row_result?;
         let perm: String = row.get(0)?;
@@ -48,44 +54,75 @@ pub fn edit_user_permissions(
     username: String,
     pool: &Pool,
     permissions: Permissions,
-) -> Result<String> {
-    let conn = pool.get()?;
+) -> Result<String, oracle::Error> {
+    let conn = pool.get();
+    if conn.is_err() {
+        error!("Error connecting to db");
+        return Err(conn.err().unwrap());
+    }
+    let conn = conn.unwrap();
+
     let user_id = username.to_string();
-    let mut stmt = conn
+    let stmt = conn
         .statement("DELETE FROM ODBC_JHC.PERMISSIONS_JHC WHERE USERNAME = :user_id")
-        .build()?;
-    stmt.execute(&[&user_id])?;
+        .build();
+    if stmt.is_err() {
+        error!("Error building statement");
+        return Err(stmt.err().unwrap());
+    }
+    let mut stmt = stmt.unwrap();
 
-    let mut stmt = conn.statement("INSERT INTO ODBC_JHC.PERMISSIONS_JHC (USERNAME, PERMISSION) VALUES (:user_id, :permission)").build()?;
 
-    if permissions.users.unwrap() {
+    match stmt.execute(&[&user_id]) {
+        Ok(_) => (),
+        Err(err) => {
+            error!("Error executing query: {}", err);
+            return Err(err);
+        }
+    };
+
+    let stmt = conn.statement("INSERT INTO ODBC_JHC.PERMISSIONS_JHC (USERNAME, PERMISSION) VALUES (:user_id, :permission)").build();
+    if stmt.is_err() {
+        error!("Error building statement");
+        return Err(stmt.err().unwrap());
+    }
+    let mut stmt = stmt.unwrap();
+
+    if permissions.users.unwrap_or(false) {
         stmt.execute(&[&user_id, &"users".to_string()])?;
     }
-    if permissions.permissions.unwrap() {
+    if permissions.permissions.unwrap_or(false) {
         stmt.execute(&[&user_id, &"permissions".to_string()])?;
     }
-    if permissions.query.unwrap() {
+    if permissions.query.unwrap_or(false) {
         stmt.execute(&[&user_id, &"query".to_string()])?;
     }
-    if permissions.images.unwrap() {
+    if permissions.images.unwrap_or(false) {
         stmt.execute(&[&user_id, &"images".to_string()])?;
     }
-    if permissions.cost.unwrap() {
+    if permissions.cost.unwrap_or(false) {
         stmt.execute(&[&user_id, &"cost".to_string()])?;
     }
-    if permissions.admin.unwrap() {
+    if permissions.admin.unwrap_or(false) {
         stmt.execute(&[&user_id, &"admin".to_string()])?;
     }
-    if permissions.stock.unwrap() {
+    if permissions.stock.unwrap_or(false) {
         stmt.execute(&[&user_id, &"stock".to_string()])?;
     }
-    if permissions.reports.unwrap() {
+    if permissions.reports.unwrap_or(false) {
         stmt.execute(&[&user_id, &"reports".to_string()])?;
     }
-    if permissions.stores.unwrap() {
+    if permissions.stores.unwrap_or(false) {
         stmt.execute(&[&user_id, &"stores".to_string()])?;
     }
 
-    conn.commit()?;
+    match conn.commit() {
+        Ok(_) => (),
+        Err(err) => {
+            error!("Error: {}", err);
+            return Err(err);
+        }
+    }
+
     Ok("Permissions Updated".to_string())
 }
